Move AcraServer common files into "common" package
This commit is intentionally broken to provide better diffs later. Only
minimum amount of changes are made: move the file, rename the package,
gofmt imports and source code.
